requirements.txt
google-cloud-bigquery>=3.0.0
requests>=2.31.0
certifi>=2023.0.0
